+++ src/common/docBuilder/docBuilderPackage.ts
@@ -2,6 +2,9 @@ import sortArray from "sort-array";
 import { PromptTemplate } from "../aiProvider/promptTemplates.js";
 import { buildGenericMarkdownTable } from "../utils/flowVisualiser/nodeFormatUtils.js";
 import { DocBuilderRoot } from "./docBuilderRoot.js";
+import { DocBuilderPackageXML } from "./docBuilderPackageXml.js";
+import fs from "fs";
+import { parsePackageXmlFile } from "../utils/xmlUtils.js";
 
 export class DocBuilderPackage extends DocBuilderRoot {
 
@@ -9,6 +12,7 @@ export class DocBuilderPackage extends DocBuilderRoot {
   public promptKey: PromptTemplate = "PROMPT_DESCRIBE_PACKAGE";
   public placeholder = "<!-- Package description -->";
   public xmlRootKey = "json";
+  public docsSection = "packages";
 
   public static buildIndexTable(prefix: string, packageDescriptions: any, filterObject: string | null = null) {
     const filteredPackages = filterObject ? packageDescriptions.filter(page => page.impactedObjects.includes(filterObject)) : packageDescriptions;
@@ -40,6 +44,19 @@ export class DocBuilderPackage extends DocBuilderRoot {
       '',
       buildGenericMarkdownTable(this.parsedXmlObject, ["SubscriberPackageName", "SubscriberPackageNamespace", "SubscriberPackageVersionNumber", "SubscriberPackageVersionId", "SubscriberPackageVersionName", "SubscriberPackageId"], "## Package attributes", []),
       '',
+      '## Package Metadatas',
+      '',
+      '<div id="jstree-container"></div>',
+      ''
     ];
+  }
+
+  // Generate json for display with jsTree npm library 
+  public async generateJsonTree(): Promise<any> {
+    if (this.additionalVariables.PACKAGE_FILE && fs.existsSync(this.additionalVariables.PACKAGE_FILE)) {
+      const packageData = await parsePackageXmlFile(this.additionalVariables.PACKAGE_FILE);
+      return DocBuilderPackageXML.generateJsonTree("all", packageData);
+    }
+    return null;
   }
 }