+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 - [hardis:doc:project2markdown](https://sfdx-hardis.cloudity.com/hardis/doc/project2markdown/):
   - Add doc for installed packages, enhanced with LLM
   - Fix markdown returned by LLMs so it is compliant with mkdocs
+  - Allow to define a property **truncateAfter** on prompts variables to avoid crashes in case value is too long
   - Authorizations doc:
     - Filter non accessible items from tree
     - Display special icons for ModifyAllData and ViewAllData items